requirements.txt
a2a-sdk==0.3.1
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.2.1
colorama==0.4.6 ; sys_platform == 'win32'
exceptiongroup==1.3.0 ; python_full_version < '3.11'
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib-metadata==8.7.0
opentelemetry-api==1.34.0
opentelemetry-sdk==1.34.0
opentelemetry-semantic-conventions==0.55b0
pydantic==2.11.5
pydantic-core==2.33.2
sniffio==1.3.1
sse-starlette==2.3.6
starlette==0.47.0
typing-extensions==4.14.0
typing-inspection==0.4.1
uvicorn==0.34.3
zipp==3.22.0